manga_web/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile manga_web/requirements/base.in
#
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via watchgod
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==20.2.0
    # via twisted
billiard==3.6.4.0
    # via celery
boto3==1.24.2
    # via -r manga_web/requirements/base.in
botocore==1.27.2
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via
    #   -r manga_web/requirements/base.in
    #   django-celery-beat
    #   django-celery-email
    #   django-celery-results
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
cryptography==37.0.2
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   scrapy
deprecated==1.2.13
    # via redis
django==4.0.5
    # via
    #   -r manga_web/requirements/base.in
    #   django-appconf
    #   django-celery-beat
    #   django-celery-email
    #   django-contact-form
    #   django-recaptcha
django-appconf==1.0.5
    # via django-celery-email
django-celery-beat==2.3.0
    # via -r manga_web/requirements/base.in
django-celery-email==3.0.0
    # via -r manga_web/requirements/base.in
django-celery-results==2.3.1
    # via -r manga_web/requirements/base.in
django-contact-form==2.0.1
    # via -r manga_web/requirements/base.in
django-environ==0.8.1
    # via -r manga_web/requirements/base.in
django-recaptcha==3.0.0
    # via -r manga_web/requirements/base.in
django-timezone-field==5.0
    # via django-celery-beat
django-widget-tweaks==1.4.12
    # via -r manga_web/requirements/base.in
feedparser==6.0.10
    # via -r manga_web/requirements/base.in
filelock==3.7.1
    # via tldextract
hyperlink==21.0.0
    # via twisted
idna==3.3
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   tldextract
incremental==21.3.0
    # via twisted
itemadapter==0.6.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.4
    # via scrapy
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
    #   itemloaders
kindlecomicconverter==5.5.2
    # via -r manga_web/requirements/base.in
kombu==5.2.4
    # via celery
lxml==4.9.0
    # via
    #   -r manga_web/requirements/base.in
    #   parsel
    #   scrapy
packaging==21.3
    # via redis
parsel==1.6.0
    # via
    #   itemloaders
    #   scrapy
pillow==9.1.1
    # via kindlecomicconverter
prompt-toolkit==3.0.29
    # via click-repl
protego==0.2.1
    # via scrapy
psutil==5.9.1
    # via kindlecomicconverter
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pydispatcher==2.0.5
    # via scrapy
pyopenssl==22.0.0
    # via scrapy
pyparsing==3.0.9
    # via packaging
pyqt5==5.15.6
    # via kindlecomicconverter
pyqt5-qt5==5.15.2
    # via pyqt5
pyqt5-sip==12.10.1
    # via pyqt5
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   python-crontab
python-slugify==6.1.2
    # via kindlecomicconverter
pytz==2022.1
    # via
    #   celery
    #   django-timezone-field
queuelib==1.6.2
    # via scrapy
raven==6.10.0
    # via kindlecomicconverter
redis==4.3.3
    # via -r manga_web/requirements/base.in
requests==2.27.1
    # via
    #   -r manga_web/requirements/base.in
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
s3transfer==0.6.0
    # via boto3
scrapy==2.6.1
    # via -r manga_web/requirements/base.in
service-identity==21.1.0
    # via scrapy
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   parsel
    #   protego
    #   python-dateutil
    #   requests-file
    #   service-identity
    #   w3lib
sniffio==1.2.0
    # via anyio
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via python-slugify
tldextract==3.3.0
    # via scrapy
twisted==22.4.0
    # via scrapy
typing-extensions==4.2.0
    # via twisted
tzdata==2022.1
    # via django-celery-beat
unidecode==1.3.4
    # via -r manga_web/requirements/base.in
urllib3==1.26.9
    # via
    #   botocore
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
w3lib==1.22.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
watchgod==0.8.2
    # via -r manga_web/requirements/base.in
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==6.1.0
    # via -r manga_web/requirements/base.in
wrapt==1.14.1
    # via deprecated
zope-interface==5.4.0
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
